catch panics in the engine, and turn them into errors
The engine ideally shouldn't panic, but it still does sometimes, and
especially during developement. We have a panic recovery mechanism,
but that doesn't work for recovering from panics from goroutines.
This is mainly annoying for two reasons:
1) When this happens to a user, we never get to relinquish the display
   control, so the terminal ends up being messed up until the user
   `reset`s it manually.
2) When running multiple tests at once, they all break, and it's
   impossible to tell which one caused the panic.  Similarly in fuzz
   testing, the fuzz tester can't recover from the panic, and thus
   can't produce a repro, because the program just exits.

For now catch the panics in deployments, and turn them into an error.
This avoids the issues described above.